+++ Prac_07/Dynamic_Button_List/people.py
